requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
django==6.0.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
django-bootstrap5==26.1
rollbar==1.3.0
django-filter==25.2
django-widget-tweaks==1.5.1
pytest-django==4.11.1
pytest-cov==6.1.1
pyyaml==6.0.2
requests==2.31
beautifulsoup4
validators